Refactoring the view engine slightly.
