+++ lib/middleware/local_auth.js
@@ -8,7 +8,8 @@ var passport = require('passport'),
     i10n = require('../i10n'),
     csurf = require('csurf'),
     Conf = require('../conf'),
-    requireHttps = require('require-https');
+    requireHttps = require('require-https'),
+    view = require('../view');
 
 var csrfProtection = csurf();
 var boundLogger = logging.getRootLogger('local_auth');
@@ -138,17 +139,10 @@ exports.passportPaths = function(app) {
 
   app.get('/\\$login/', csrfProtection, function(req, res) {
     res.cacheControl.noCache();
-    var view = {meta: {modified: null, created:null}, user: req.user};
-    view.errorMessage = req.getFlashMsgs('error');
-    view.infoMessage = req.getFlashMsgs('info');
-    view.intl = i10n.getIntl();
-    view.site = req.site;
-    view.scheme = req.scheme;
-    view.security = req.access;
-    view.csrfToken = req.csrfToken();
-    res.expose(view.intl, 'intl');
-    view.expose = res.locals.state.toString();
-    req.scheme.render('login', view,
+    var data = view.basicViewSetup(req, res);
+    data.csrfToken = req.csrfToken();
+    res.expose(data.intl, 'intl');
+    req.scheme.render('login', data,
       function(err, page) {
         if (err) {
           res.write("ERROR");
