+++ lib/view.js
@@ -60,6 +60,18 @@ function noCacheView(req, res, next) {
   next();
 }
 
+function basicViewSetup(req, res) {
+  var view = {meta: {modified: null, created:null}, user: req.user};
+  view.errorMessage = req.getFlashMsgs('error');
+  view.infoMessage = req.getFlashMsgs('info');
+  view.intl = i10n.getIntl();
+  view.site = req.site;
+  view.scheme = req.scheme;
+  view.security = req.access;
+  res.expose(view.intl, 'intl');
+  return view;
+};
+
 function basicView(section, template, req, res, next) {
   var view = req.entity.view();
   view.user = undefined;
@@ -123,6 +135,7 @@ function basicViewWithJsonView(section, template, req, res, next) {
   });
 }
 
+exports.basicViewSetup = basicViewSetup;
 exports.basicView = basicView;
 exports.noCacheView = noCacheView;
 exports.basicViewWithJsonView = basicViewWithJsonView;
